rename mass deployer to tfrobot
