+++ .github/workflows/lint.yml
@@ -36,4 +36,4 @@ jobs:
           - monitoring-bot
           - rmb-sdk-go
           - user-contracts-mon 
-          - mass-deployer
+          - tfrobot
