+++ .github/workflows/test.yml
@@ -40,4 +40,4 @@ jobs:
           - monitoring-bot
           - rmb-sdk-go
           - user-contracts-mon 
-          - mass-deployer
+          - tfrobot
