--- mass-deployer/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/threefoldtech/tfgrid-sdk-go/mass-deployer/cmd"
-
-func main() {
-	cmd.Execute()
-}
