+++ tfrobot/cmd/cancel.go
@@ -2,32 +2,23 @@
 package cmd
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
 
 	"github.com/rs/zerolog"
 	"github.com/spf13/cobra"
-	"github.com/threefoldtech/tfgrid-sdk-go/mass-deployer/internal/parser"
-	deployer "github.com/threefoldtech/tfgrid-sdk-go/mass-deployer/pkg/mass-deployer"
-	"golang.org/x/sys/unix"
+	"github.com/threefoldtech/tfgrid-sdk-go/tfrobot/internal/parser"
+	"github.com/threefoldtech/tfgrid-sdk-go/tfrobot/pkg/deployer"
 )
 
-const (
-	ymlExt  = ".yml"
-	yamlExt = ".yaml"
-	jsonExt = ".json"
-)
-
-var deployCmd = &cobra.Command{
-	Use:   "deploy",
-	Short: "deploy groups of vms in configuration file",
+var cancelCmd = &cobra.Command{
+	Use:   "cancel",
+	Short: "cancel all deployments of configuration file",
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// It doesn't have a subcommand
 		if len(cmd.Flags().Args()) != 0 {
-			return fmt.Errorf("'deploy' and %v cannot be used together, please use one command at a time", cmd.Flags().Args())
+			return fmt.Errorf("'cancel' and %v cannot be used together, please use one command at a time", cmd.Flags().Args())
 		}
 
 		debug, err := cmd.Flags().GetBool("debug")
@@ -38,23 +29,6 @@ var deployCmd = &cobra.Command{
 		zerolog.SetGlobalLevel(zerolog.InfoLevel)
 		if debug {
 			zerolog.SetGlobalLevel(zerolog.DebugLevel)
-		}
-
-		outputPath, err := cmd.Flags().GetString("output")
-		if err != nil {
-			return fmt.Errorf("error in output file: %w", err)
-		}
-
-		outJsonFmt := filepath.Ext(outputPath) == jsonExt
-		outYmlFmt := filepath.Ext(outputPath) == yamlExt || filepath.Ext(outputPath) == ymlExt
-		if !outJsonFmt && !outYmlFmt {
-			return fmt.Errorf("unsupported output file format '%s', should be [yaml, yml, json]", outputPath)
-		}
-
-		_, err = os.Stat(outputPath)
-		// check if output file is writable
-		if !errors.Is(err, os.ErrNotExist) && unix.Access(outputPath, unix.W_OK) != nil {
-			return fmt.Errorf("output path '%s' is not writable", outputPath)
 		}
 
 		configPath, err := cmd.Flags().GetString("config")
@@ -88,13 +62,9 @@ var deployCmd = &cobra.Command{
 			return err
 		}
 
-		if err = parser.ValidateConfig(cfg, tfPluginClient); err != nil {
-			return fmt.Errorf("failed to validate configuration file '%s' with error: %w", configPath, err)
-		}
-
-		ctx := context.Background()
-		if err = deployer.RunDeployer(ctx, cfg, tfPluginClient, outputPath, debug); err != nil {
-			return fmt.Errorf("failed to run the deployer with error: %w", err)
+		err = deployer.RunCanceler(cfg, tfPluginClient, debug)
+		if err != nil {
+			return fmt.Errorf("failed to cancel configured deployments with error: %w", err)
 		}
 
 		return nil
