+++ tfrobot/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/threefoldtech/tfgrid-sdk-go/tfrobot/cmd"
+
+func main() {
+	cmd.Execute()
+}
